Modify com.sigio.util.ValueTransformer and family to use generics.
ValueTransformer and its subclasses now use generics for the type
of the object whose values they will transform.  This eliminates the
need for type checking and throwing of IllegalArgumentException.

NumberToDoubleValueTransformer was deleted as it is no longer needed.

I may add an ObjectTransformer class in the future to handle cases
where one would want to transforme one object type to another.

Signed-off-by: Jason Stephenson <jason@sigio.com>